feat: wrap components in Suspense for improved loading handling
