+++ src/app/admin/emulators/[emulatorId]/page.tsx
@@ -1,126 +1,20 @@
-'use client'
+import { Suspense } from 'react'
+import EditEmulatorClientPage from './EditEmulatorClientPage'
 
-import { ArrowLeft } from 'lucide-react'
-import { notFound, useParams, useRouter } from 'next/navigation'
-import { Button, PageSkeletonLoading } from '@/components/ui'
-import { api } from '@/lib/api'
-import { hasPermission, PERMISSIONS } from '@/utils/permission-system'
-import { Role } from '@orm'
-import EmulatorEditForm from './components/EmulatorEditForm'
-import ManageSupportedSystems from './components/ManageSupportedSystems'
+interface Props {
+  params: Promise<{ emulatorId: string }>
+}
 
-function EditEmulatorPage() {
-  const router = useRouter()
-  const params = useParams()
-  const emulatorId = params.emulatorId as string
-
-  const emulatorsQuery = api.emulators.byId.useQuery({ id: emulatorId }, { enabled: !!emulatorId })
-  const currentUser = api.users.me.useQuery()
-  const verifiedDeveloperQuery = api.emulators.getVerifiedDeveloper.useQuery(
-    { emulatorId },
-    {
-      enabled: !!emulatorId && currentUser.data?.role === Role.DEVELOPER,
-    },
-  )
-
-  const user = currentUser.data
-  const hasManagePermission = hasPermission(user?.permissions, PERMISSIONS.MANAGE_CUSTOM_FIELDS)
-  const isDeveloper = user?.role === Role.DEVELOPER
-  const isVerifiedDeveloper = Boolean(verifiedDeveloperQuery.data)
-  const hasAccess = Boolean(user && hasManagePermission && (!isDeveloper || isVerifiedDeveloper))
-
-  const systemsQuery = api.systems.get.useQuery({}, { enabled: !!user && hasAccess })
-
-  const isLoading =
-    currentUser.isPending || emulatorsQuery.isLoading || verifiedDeveloperQuery.isLoading
-
-  if (isLoading) return <PageSkeletonLoading />
-
-  if (!user || !hasAccess) {
-    return (
-      <div className="container mx-auto p-8">
-        <Button
-          variant="outline"
-          size="sm"
-          icon={ArrowLeft}
-          onClick={() => router.back()}
-          className="mb-6"
-        >
-          Back
-        </Button>
-        <p className="text-lg text-gray-700 dark:text-gray-200">
-          You do not have permission to manage this emulator.
-        </p>
-      </div>
-    )
-  }
-
-  if (systemsQuery.isPending) return <PageSkeletonLoading />
-
-  if (emulatorsQuery.error || systemsQuery.error) {
-    return (
-      <div className="container mx-auto p-4">
-        <p className="text-red-500">
-          Error loading data:{' '}
-          {emulatorsQuery.error?.message ??
-            systemsQuery.error?.message ??
-            'An unknown error occurred'}
-        </p>
-      </div>
-    )
-  }
-
-  if (!emulatorsQuery.data || !systemsQuery.data) return notFound()
-
+export default function EditEmulatorPage(props: Props) {
   return (
-    <div className="container mx-auto p-4 md:p-8">
-      <Button
-        variant="outline"
-        size="sm"
-        icon={ArrowLeft}
-        onClick={() => router.back()}
-        className="mb-6"
-      >
-        Back
-      </Button>
-
-      <h1 className="text-3xl font-bold mb-8 text-gray-800 dark:text-white">
-        Edit Emulator:{' '}
-        <span className="text-blue-600 dark:text-blue-400">{emulatorsQuery.data.name}</span>
-      </h1>
-
-      <div className="space-y-10">
-        {/* Section for Emulator Details */}
-        <section className="bg-white dark:bg-gray-800 shadow-xl rounded-lg p-6">
-          <h2 className="text-xl font-semibold mb-6 text-gray-700 dark:text-gray-200 border-b pb-3 dark:border-gray-700">
-            Emulator Details
-          </h2>
-          <EmulatorEditForm emulator={emulatorsQuery.data} />
-        </section>
-
-        {/* Section for Supported Systems */}
-        <section className="bg-white dark:bg-gray-800 shadow-xl rounded-lg p-6">
-          <h2 className="text-xl font-semibold mb-6 text-gray-700 dark:text-gray-200 border-b pb-3 dark:border-gray-700">
-            Supported Systems
-          </h2>
-          <ManageSupportedSystems
-            emulatorId={emulatorsQuery.data.id}
-            allSystems={systemsQuery.data}
-            currentlySupportedSystems={emulatorsQuery.data.systems} // emulator.systems should be populated by the byId query
-          />
-        </section>
-
-        <section className="bg-white dark:bg-gray-800 shadow-xl rounded-lg p-6">
-          <h2 className="text-xl font-semibold mb-6 text-gray-700 dark:text-gray-200 border-b pb-3 dark:border-gray-700">
-            Custom Fields
-          </h2>
-          <Button onClick={() => router.push(`/admin/emulators/${emulatorId}/custom-fields`)}>
-            Manage Custom Fields
-          </Button>
-        </section>
-      </div>
-    </div>
+    <Suspense fallback={<div className="container mx-auto p-8">Loading emulator...</div>}>
+      <EditEmulatorPageContent params={props.params} />
+    </Suspense>
   )
 }
 
-export default EditEmulatorPage
+async function EditEmulatorPageContent(props: Props) {
+  const { emulatorId } = await props.params
+
+  return <EditEmulatorClientPage emulatorId={emulatorId} />
+}
